Refactored Alcohol List to render individual drinks so that the Rodal works separately for each drink
